requirements.txt
Flask==3.0.3
psutil==6.0.0
python-dotenv==1.0.1
